refactor: 일정 년도별 조회 수정
