+++ src/App.jsx
@@ -36,7 +36,6 @@ import { BoardProvider } from '@/api/BoardContext';
 import { DuesProvider } from '@/api/DuesContext';
 import { NoticeProvider } from '@/api/NoticeContext';
 import { UserProvider } from '@/api/UserContext';
-import { YearlyScheduleProvider } from '@/api/YearlyScheduleContext';
 
 import ScrollToTop from '@/hooks/ScrollToTop';
 
@@ -47,35 +46,33 @@ const App = () => {
         <BoardProvider>
           <DuesProvider>
             <NoticeProvider>
-              <YearlyScheduleProvider>
-                <AttendCheckProvider>
-                  <ScrollToTop />
-                  <Routes>
-                    <Route path="/" element={<Landing />} />
-                    <Route path="/login" element={<Login />} />
-                    <Route path="/attendance" element={<Attendance />} />
-                    <Route path="/calendar" element={<Calendar />} />
-                    <Route path="/:type/:id" element={<EventDetails />} />
-                    <Route path="/event/create" element={<EventAdmin />} />
-                    <Route path="/event/:id/edit" element={<EventAdmin />} />
-                    <Route path="/home" element={<Home />} />
-                    <Route path="/signup" element={<Signup />} />
-                    <Route path="/profile" element={<Profile />} />
-                    <Route path="/attendCheck" element={<AttendCheck />} />
-                    <Route path="/member" element={<Member />} />
-                    <Route path="/member/:id" element={<MemberDetail />} />
-                    <Route path="/mypage" element={<MyPage />} />
-                    <Route path="/edit" element={<Edit />} />
-                    <Route path="/dues" element={<Dues />} />
-                    <Route path="/receipt" element={<Receipt />} />
-                    <Route path="/board" element={<Board />} />
-                    <Route path="/study/:id" element={<StudyDetail />} />
-                    <Route path="/notice/:id" element={<NoticeDetail />} />
-                    <Route path="/study/post" element={<StudyPosting />} />
-                    <Route path="/notice/post" element={<NoticePosting />} />
-                  </Routes>
-                </AttendCheckProvider>
-              </YearlyScheduleProvider>
+              <AttendCheckProvider>
+                <ScrollToTop />
+                <Routes>
+                  <Route path="/" element={<Landing />} />
+                  <Route path="/login" element={<Login />} />
+                  <Route path="/attendance" element={<Attendance />} />
+                  <Route path="/calendar" element={<Calendar />} />
+                  <Route path="/:type/:id" element={<EventDetails />} />
+                  <Route path="/event/create" element={<EventAdmin />} />
+                  <Route path="/event/:id/edit" element={<EventAdmin />} />
+                  <Route path="/home" element={<Home />} />
+                  <Route path="/signup" element={<Signup />} />
+                  <Route path="/profile" element={<Profile />} />
+                  <Route path="/attendCheck" element={<AttendCheck />} />
+                  <Route path="/member" element={<Member />} />
+                  <Route path="/member/:id" element={<MemberDetail />} />
+                  <Route path="/mypage" element={<MyPage />} />
+                  <Route path="/edit" element={<Edit />} />
+                  <Route path="/dues" element={<Dues />} />
+                  <Route path="/receipt" element={<Receipt />} />
+                  <Route path="/board" element={<Board />} />
+                  <Route path="/study/:id" element={<StudyDetail />} />
+                  <Route path="/notice/:id" element={<NoticeDetail />} />
+                  <Route path="/study/post" element={<StudyPosting />} />
+                  <Route path="/notice/post" element={<NoticePosting />} />
+                </Routes>
+              </AttendCheckProvider>
             </NoticeProvider>
           </DuesProvider>
         </BoardProvider>
@@ -85,4 +82,3 @@ const App = () => {
 };
 
 export default App;
-
