+++ src/components/Calendar/YearCalendar.tsx
@@ -1,47 +1,61 @@
-import * as S from '@/styles/calendar/YearCalendar.styled';
-import { useContext } from 'react';
-
+import getYearlySchedule from '@/api/YearlyScheduleAPI';
 import { MONTH } from '@/constants/dateConstants';
-import { YearlyScheduleContext } from '@/api/YearlyScheduleContext';
+import * as S from '@/styles/calendar/YearCalendar.styled';
+import { useEffect, useState } from 'react';
 import YearlyCard from './YearlyCard';
 
 const YearCalendar = ({ year }: { year: string }) => {
   const validYear = year.toString().length === 4 ? parseInt(year, 10) : 2024;
-  const { yearScheduleData, error } = useContext(YearlyScheduleContext);
 
-  if (error) {
-    return <S.Error>데이터를 불러오는 중 문제가 발생했습니다</S.Error>;
-  }
+  // TODO: 서버 수정 후 자세한 타입 지정
+  const [yearlySchedule, setYearlySchedule] = useState<Record<number, any>>({});
 
-  if (!yearScheduleData) {
-    return <S.Error>Loading...</S.Error>;
-  }
+  useEffect(() => {
+    const fetchData = async () => {
+      try {
+        if (year) {
+          const response = await getYearlySchedule(
+            `${year}-01-01T00:00:00.000Z`,
+            `${year}-12-31T23:59:59.999Z`,
+          );
+          if (response.data.code === 200) {
+            setYearlySchedule(response.data.data);
+          } else {
+            console.error(response.data.message);
+          }
+        }
+      } catch (err) {
+        console.error(err);
+      }
+    };
+    fetchData();
+  }, [year]);
 
   return (
     <S.MonthlyBox>
       <S.FirstHalfMonth>
-        {MONTH.filter((monthItem) => monthItem >= 1 && monthItem <= 6).map(
-          (monthItem) => (
-            <YearlyCard
-              key={monthItem}
-              thisMonth={monthItem}
-              year={validYear}
-              events={yearScheduleData[monthItem] || []}
-            />
-          ),
-        )}
+        {MONTH.filter(
+          (monthItem) => yearlySchedule && monthItem >= 1 && monthItem <= 6,
+        ).map((monthItem) => (
+          <YearlyCard
+            key={monthItem}
+            thisMonth={monthItem}
+            year={validYear}
+            events={yearlySchedule[monthItem] || []}
+          />
+        ))}
       </S.FirstHalfMonth>
       <S.SecondHalfMonth>
-        {MONTH.filter((monthItem) => monthItem >= 7 && monthItem <= 12).map(
-          (monthItem) => (
-            <YearlyCard
-              key={monthItem}
-              thisMonth={monthItem}
-              year={validYear}
-              events={yearScheduleData[monthItem] || []}
-            />
-          ),
-        )}
+        {MONTH.filter(
+          (monthItem) => yearlySchedule && monthItem >= 7 && monthItem <= 12,
+        ).map((monthItem) => (
+          <YearlyCard
+            key={monthItem}
+            thisMonth={monthItem}
+            year={validYear}
+            events={yearlySchedule[monthItem] || []}
+          />
+        ))}
       </S.SecondHalfMonth>
     </S.MonthlyBox>
   );
