+++ src/pages/Calendar.tsx
@@ -6,7 +6,6 @@ import ToggleButton from '@/components/Calendar/ToggleButton';
 import YearCalendar from '@/components/Calendar/YearCalendar';
 import { CURRENT_MONTH, CURRENT_YEAR } from '@/constants/dateConstants';
 import useCustomBack from '@/hooks/useCustomBack';
-import YearlyScheduleAPI from '@/api/YearlyScheduleAPI';
 
 import * as S from '@/styles/calendar/Calendar.styled';
 
@@ -16,17 +15,12 @@ const Calendar = () => {
   const [year, setYear] = useState(CURRENT_YEAR);
   const [month, setMonth] = useState(CURRENT_MONTH);
 
-  // TODO: 백엔드 수정 후 삭제 예정
-  const start = `${year}-01-01T00:00:00.000Z`;
-  const end = `${year}-12-31T23:59:59.999Z`;
-
   const onToggle = () => {
     setIsMonth(!isMonth);
   };
 
   return (
     <S.CalendarWrapper>
-      <YearlyScheduleAPI start={start} end={end} />
       <CalendarHeader
         month={month}
         year={year}
